requirements.txt
ipython
django
djangorestframework
django-cors-headers
uvicorn
google-generativeai
google-api-python-client 
google-auth-httplib2
google-auth-oauthlib
pymongo
pymongo[srv]==3.12

Crawl4AI
python-dotenv
certifi
pyaml
google-generativeai
openai
supabase

langchain
langgraph
langgraph-supervisor
langsmith
langchain_anthropic
langchain-google-genai
streamlit
llama-index
llama-index-llms-gemini
llama-index-embeddings-gemini
llama-parse
llama-index-extractors-entity
llama-index-postprocessor-cohere-rerank
fs

faiss-cpu